requirements.txt
fastapi
uvicorn
pydantic
aiofiles
python-dotenv
openai
python-multipart
requests
rapidfuzz
numpy
scipy
tqdm
timeout-decorator
google-generativeai
litellm
httpx
aiohttp